+++ Tankman Project/Assets/Scripts/BotsManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25cdebe191464954dbb21a982f62435a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
