+++ Tankman Project/Assets/Scripts/Map/MapsManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5da640d1bfc8ecd448e65e6fa903bb8c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
