+++ Tankman Project/Assets/_TerrainAutoUpgrade.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9d11bf271492e254d9d87b71559a74dd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
