+++ Tankman Project/Assets/_TerrainAutoUpgrade/layer_UnitTextureBluecc0f254a2c535401.terrainlayer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0d6069f0240f81743b3db346191ff028
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
