+++ Tankman Project/Assets/_TerrainAutoUpgrade/layer_UnitTextureRedcc0f254a2c535401.terrainlayer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0788117f8da6b9b4b86aa05c3e53e5fc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
