+++ Tankman Project/Logs/Packages-Update.log
@@ -1,0 +1,25 @@
+
+=== Fri Jul 17 23:52:15 2020
+
+Packages were changed.
+Update Mode: updateDependencies
+
+The following packages were added:
+  com.unity.2d.sprite@1.0.0
+  com.unity.2d.tilemap@1.0.0
+  com.unity.collab-proxy@1.2.16
+  com.unity.ide.rider@1.1.4
+  com.unity.ide.vscode@1.2.1
+  com.unity.modules.androidjni@1.0.0
+  com.unity.multiplayer-hlapi@1.0.6
+  com.unity.test-framework@1.1.14
+  com.unity.timeline@1.2.6
+  com.unity.ugui@1.0.0
+  com.unity.xr.legacyinputhelpers@2.1.4
+The following packages were updated:
+  com.unity.ads from version 2.0.8 to 3.4.7
+  com.unity.analytics from version 2.0.16 to 3.3.5
+  com.unity.purchasing from version 2.0.3 to 2.0.6
+  com.unity.textmeshpro from version 1.2.4 to 2.0.1
+The following packages were removed:
+  com.unity.package-manager-ui@1.9.11
