+++ Tankman Project/ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,2 @@
-m_EditorVersion: 2018.2.20f1
+m_EditorVersion: 2019.4.4f1
+m_EditorVersionWithRevision: 2019.4.4f1 (1f1dac67805b)
